{"schema":"decodex/commit/1","summary":"Modularize runtime and MCP test batches","authority":"manual"}
