+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/lifecycle/protocol_activity/attempt_scoping.rs
@@ -1,0 +1,47 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, OffsetDateTime, StateStore, fs, orchestrator, process, state,
+};
+
+#[test]
+fn operator_status_snapshot_ignores_marker_from_newer_attempt_for_stored_run() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "failed")
+		.expect("stored run should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+	state::write_run_activity_marker_for_process(&worktree_path, "run-2", 2, process::id())
+		.expect("newer attempt marker should write");
+	state::write_run_retry_schedule(
+		&worktree_path,
+		"run-2",
+		2,
+		"failure",
+		OffsetDateTime::now_utc().unix_timestamp() + 60,
+	)
+	.expect("retry schedule marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.recent_runs.first().expect("recent run should exist");
+
+	assert_eq!(run.run_id, "run-1");
+	assert_eq!(run.phase, "failed");
+	assert_eq!(run.wait_reason, None);
+	assert_eq!(run.process_id, None);
+	assert_eq!(run.process_alive, None);
+	assert_eq!(run.retry_kind, None);
+	assert_eq!(run.next_retry_at, None);
+}
