+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/lifecycle/protocol_activity/sanitization.rs
@@ -1,0 +1,74 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, ProtocolActivityMarker, ProtocolActivitySummary, StateStore, orchestrator, state,
+};
+
+#[test]
+fn operator_status_snapshot_sanitizes_private_protocol_activity_details() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+	let protocol_activity = ProtocolActivitySummary {
+		turn_status: Some(String::from("running")),
+		waiting_reason: Some(String::from("tool_execution")),
+		rate_limit_status: None,
+		recent_events: vec![
+			state::ProtocolActivityEventSummary {
+				event_type: String::from("configWarning"),
+				category: String::from("warning"),
+				detail: Some(String::from("state marker path=/srv/decodex/runtime")),
+			},
+			state::ProtocolActivityEventSummary {
+				event_type: String::from("configWarning"),
+				category: String::from("warning"),
+				detail: Some(String::from("state marker (/srv/decodex/runtime)")),
+			},
+		],
+	};
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 2,
+			last_event_type: "configWarning",
+			child_agent_activity: None,
+			protocol_activity: Some(&protocol_activity),
+		},
+	)
+	.expect("protocol activity marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should exist");
+	let rendered = orchestrator::render_operator_status(&snapshot);
+	let summary = run.protocol_activity.as_ref().expect("protocol summary should render");
+
+	assert!(
+		summary
+			.recent_events
+			.iter()
+			.all(|event| event.detail.as_deref() == Some("redacted_sensitive_detail"))
+	);
+	assert!(rendered.contains("configWarning:redacted_sensitive_detail"));
+	assert!(!rendered.contains("path=/srv"));
+	assert!(!rendered.contains("(/srv"));
+}
