+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/lifecycle/protocol_activity/summary_projection.rs
@@ -1,0 +1,119 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, ProtocolActivityMarker, ProtocolActivitySummary, StateStore, orchestrator, state,
+};
+
+#[test]
+fn operator_status_snapshot_uses_structured_protocol_activity_summary() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+	let protocol_activity = ProtocolActivitySummary {
+		turn_status: Some(String::from("running")),
+		waiting_reason: Some(String::from("approval_or_user_input")),
+		rate_limit_status: Some(String::from("primary")),
+		recent_events: vec![
+			state::ProtocolActivityEventSummary {
+				event_type: String::from("plan/update"),
+				category: String::from("plan"),
+				detail: Some(String::from("verify")),
+			},
+			state::ProtocolActivityEventSummary {
+				event_type: String::from("item/tool/requestUserInput"),
+				category: String::from("item"),
+				detail: None,
+			},
+		],
+	};
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 2,
+			last_event_type: "item/tool/requestUserInput",
+			child_agent_activity: None,
+			protocol_activity: Some(&protocol_activity),
+		},
+	)
+	.expect("protocol activity marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should exist");
+	let rendered = orchestrator::render_operator_status(&snapshot);
+
+	assert_eq!(run.wait_reason.as_deref(), Some("approval_or_user_input"));
+	assert_eq!(run.protocol_activity.as_ref(), Some(&protocol_activity));
+	assert_eq!(
+		snapshot.projects[0].waiting_lane_count, 1,
+		"approval or user-input waits should remain project-level waiting"
+	);
+	assert!(rendered.contains("protocol_activity: turn=running; waiting=approval_or_user_input; rate_limit=primary; recent=item/tool/requestUserInput, plan/update:verify"));
+}
+
+#[test]
+fn operator_status_snapshot_prefers_newer_protocol_marker_over_stale_archive_event() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+	state_store
+		.append_event("run-1", 1, "thread/archive/discarded", "{}")
+		.expect("archive event should record");
+
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 2,
+			last_event_type: "item/tool/call",
+			child_agent_activity: None,
+			protocol_activity: None,
+		},
+	)
+	.expect("protocol activity marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should exist");
+
+	assert_eq!(run.last_event_type.as_deref(), Some("item/tool/call"));
+	assert_eq!(run.event_count, 2);
+}
