+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/liveness/process_liveness/thread_status.rs
@@ -1,0 +1,138 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, ProtocolActivityMarker, StateStore, fs, orchestrator, state,
+};
+
+#[test]
+fn operator_status_snapshot_treats_dead_leased_app_server_run_as_attention_not_running() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+	state::write_run_activity_marker_for_process(&worktree_path, "run-1", 1, u32::MAX)
+		.expect("stopped process marker should write");
+	state::write_run_thread_status_marker(
+		&worktree_path,
+		"run-1",
+		1,
+		Some("thread-1"),
+		Some("turn-1"),
+		"active",
+		&[],
+	)
+	.expect("thread status should write");
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 2,
+			last_event_type: "model/response",
+			child_agent_activity: None,
+			protocol_activity: None,
+		},
+	)
+	.expect("protocol summary should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should remain visible");
+	let project = snapshot.projects.first().expect("project summary should exist");
+
+	assert_eq!(run.status, "running");
+	assert_eq!(run.attempt_status, "running");
+	assert!(run.run_lease);
+	assert_eq!(run.queue_lease_state, "held");
+	assert_eq!(run.process_alive, Some(false));
+	assert_eq!(run.process_liveness_reason.as_deref(), Some("process_stopped"));
+	assert_eq!(run.thread_status.as_deref(), Some("active"));
+	assert_eq!(run.liveness_state, "not_running");
+	assert!(run.has_fresh_execution);
+	assert!(!run.counts_as_running);
+	assert!(run.needs_attention);
+	assert_eq!(project.current_lane_count, 1);
+	assert_eq!(project.running_lane_count, 0);
+	assert_eq!(project.attention_count, 1);
+}
+
+#[test]
+fn operator_status_snapshot_treats_system_error_thread_as_attention_not_running() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+	state::write_run_thread_status_marker(
+		&worktree_path,
+		"run-1",
+		1,
+		Some("thread-1"),
+		Some("turn-1"),
+		"systemError",
+		&[],
+	)
+	.expect("thread status should write");
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 2,
+			last_event_type: "thread/status/changed",
+			child_agent_activity: None,
+			protocol_activity: None,
+		},
+	)
+	.expect("protocol summary should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should remain visible");
+	let project = snapshot.projects.first().expect("project summary should exist");
+
+	assert_eq!(run.status, "running");
+	assert_eq!(run.phase, "executing");
+	assert_eq!(run.thread_status.as_deref(), Some("systemError"));
+	assert_eq!(run.liveness_state, "process_alive");
+	assert!(run.needs_attention);
+	assert!(!run.counts_as_running);
+	assert_eq!(project.current_lane_count, 1);
+	assert_eq!(project.running_lane_count, 0);
+	assert_eq!(project.attention_count, 1);
+}
