+++ apps/decodex/src/orchestrator/tests/recovery_runtime_reentry/terminal_closeout.rs
@@ -1,342 +1,3 @@
-use std::fs;
-
-use crate::{
-	orchestrator::{
-		self, TERMINAL_GUARD_MARKER_FILE,
-		tests::{self, FakeTracker, TEST_SERVICE_ID, recovery_terminal_support},
-	},
-	state::{self, StateStore},
-	tracker,
-	worktree::WorktreeManager,
-};
-
-#[test]
-fn run_project_once_skips_recovered_terminal_guarded_worktree_after_empty_state_startup() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let issue = recovery_terminal_support::sample_active_issue_without_needs_attention_team_label(
-		"In Progress",
-	);
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("recovered worktree should exist");
-
-	fs::write(
-		worktree.path.join(TERMINAL_GUARD_MARKER_FILE),
-		"run_id=pub-101-attempt-1-123\nattempt_number=1\n",
-	)
-	.expect("terminal guard marker should write");
-
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, true)
-		.expect("recovery should succeed");
-
-	assert!(
-		summary.is_none(),
-		"restart recovery should not redispatch retained lanes guarded by a terminal marker"
-	);
-	assert!(
-		state_store
-			.worktree_for_issue(&issue.id)
-			.expect("worktree lookup should succeed")
-			.is_some(),
-		"worktree mapping should still be reconstructed for guarded retained lanes"
-	);
-}
-
-#[test]
-fn run_project_once_clears_terminal_queued_lane_labels_without_dispatch() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let active_label = tracker::automation_active_label(TEST_SERVICE_ID);
-	let issue = tests::sample_issue("Done", &[active_label.as_str()]);
-	let tracker = FakeTracker::new(vec![issue.clone()]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, false)
-		.expect("terminal queued cleanup should succeed");
-
-	assert!(summary.is_none(), "terminal queued issues should not dispatch");
-	assert_eq!(
-		tracker.label_removals.borrow().as_slice(),
-		[
-			(issue.id.clone(), vec![String::from("label-active")]),
-			(issue.id.clone(), vec![String::from("label-queued")]),
-		]
-	);
-}
-
-#[test]
-fn run_project_once_dry_run_keeps_terminal_queued_lane_labels() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let active_label = tracker::automation_active_label(TEST_SERVICE_ID);
-	let issue = tests::sample_issue("Done", &[active_label.as_str()]);
-	let tracker = FakeTracker::new(vec![issue.clone()]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, true)
-		.expect("terminal queued dry run should succeed");
-
-	assert!(summary.is_none(), "terminal queued dry run should not dispatch");
-	assert!(
-		tracker.label_removals.borrow().is_empty(),
-		"dry run should not mutate terminal queued labels"
-	);
-}
-
-#[test]
-fn run_project_once_preserves_terminal_recovered_worktree_without_prior_state_when_review_handoff_is_missing()
- {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let issue = recovery_terminal_support::sample_active_issue("Done");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("terminal retained worktree should be created");
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, false)
-		.expect("reconciliation should finish cleanly");
-
-	assert!(
-		summary.is_none(),
-		"blocked retained closeout with missing handoff should not redispatch during recovery"
-	);
-	assert!(
-		worktree.path.exists(),
-		"terminal recovery should preserve the retained closeout worktree on disk for manual intervention"
-	);
-	assert!(
-		state_store
-			.worktree_for_issue(&issue.id)
-			.expect("worktree lookup should succeed")
-			.is_some(),
-		"terminal recovery should preserve the retained closeout worktree mapping when review handoff is missing"
-	);
-}
-
-#[test]
-fn run_project_once_clears_stale_completed_closeout_lease_but_keeps_worktree() {
-	let (_temp_dir, base_config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(
-		&base_config,
-		"DECODEX_TEST_MISSING_STARTUP_DELIVERY_CLOSEOUT_GITHUB_TOKEN",
-	);
-	let issue = recovery_terminal_support::sample_active_issue("Done");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("retained closeout worktree should exist");
-	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
-	let run_id = "run-closeout-startup";
-	let pr_url = "https://github.com/hack-ink/decodex/pull/178";
-
-	tests::seed_review_handoff_marker_value(
-		&state_store,
-		config.service_id(),
-		&issue.id,
-		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
-	);
-
-	state_store
-		.record_run_attempt(run_id, &issue.id, 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease(config.service_id(), &issue.id, run_id, &issue.state.name)
-		.expect("stale lease should record");
-	state_store
-		.upsert_worktree(
-			config.service_id(),
-			&issue.id,
-			&worktree.branch_name,
-			&worktree.path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, false)
-		.expect("startup reconciliation should succeed");
-
-	assert!(
-		summary.is_none(),
-		"blocked retained closeout should not redispatch during startup recovery"
-	);
-	assert!(
-		state_store.lease_for_issue(&issue.id).expect("lease lookup should succeed").is_none(),
-		"startup reconciliation should clear stale completed closeout leases"
-	);
-	assert_eq!(
-		state_store
-			.run_attempt(run_id)
-			.expect("run attempt lookup should succeed")
-			.expect("run attempt should still exist")
-			.status(),
-		"interrupted"
-	);
-	assert!(
-		state_store
-			.worktree_for_issue(&issue.id)
-			.expect("worktree lookup should succeed")
-			.is_some(),
-		"startup reconciliation should preserve the retained closeout worktree mapping"
-	);
-	assert!(
-		worktree.path.exists(),
-		"startup reconciliation should leave the retained closeout worktree on disk"
-	);
-}
-
-#[test]
-fn run_project_once_preserves_fresh_completed_closeout_lease() {
-	let (_temp_dir, base_config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(
-		&base_config,
-		"DECODEX_TEST_MISSING_STARTUP_DELIVERY_CLOSEOUT_GITHUB_TOKEN",
-	);
-	let issue = recovery_terminal_support::sample_active_issue("Done");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("retained closeout worktree should exist");
-	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
-	let run_id = "run-closeout-fresh-startup";
-	let pr_url = "https://github.com/hack-ink/decodex/pull/178";
-
-	tests::seed_review_handoff_marker_value(
-		&state_store,
-		config.service_id(),
-		&issue.id,
-		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
-	);
-	state::write_run_activity_marker(&worktree.path, run_id, 1)
-		.expect("fresh activity marker should write");
-
-	state_store
-		.record_run_attempt(run_id, &issue.id, 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease(config.service_id(), &issue.id, run_id, &issue.state.name)
-		.expect("fresh lease should record");
-	state_store
-		.upsert_worktree(
-			config.service_id(),
-			&issue.id,
-			&worktree.branch_name,
-			&worktree.path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, false)
-		.expect("startup reconciliation should succeed");
-
-	assert!(
-		summary.is_none(),
-		"fresh retained closeout activity should block redispatch during startup recovery"
-	);
-	assert_eq!(
-		state_store
-			.lease_for_issue(&issue.id)
-			.expect("lease lookup should succeed")
-			.expect("fresh retained closeout lease should survive")
-			.run_id(),
-		run_id
-	);
-	assert_eq!(
-		state_store
-			.run_attempt(run_id)
-			.expect("run attempt lookup should succeed")
-			.expect("run attempt should still exist")
-			.status(),
-		"running"
-	);
-	assert!(
-		state_store
-			.worktree_for_issue(&issue.id)
-			.expect("worktree lookup should succeed")
-			.is_some(),
-		"startup reconciliation should preserve the retained closeout worktree mapping"
-	);
-}
-
-#[test]
-fn run_project_once_preserves_completed_unmerged_closeout_worktree() {
-	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
-	let issue = recovery_terminal_support::sample_active_issue("Done");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("retained closeout worktree should exist");
-	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
-	let run_id = "run-closeout-open-pr-startup";
-	let pr_url = "https://github.com/hack-ink/decodex/pull/179";
-	let _path_guard = recovery_terminal_support::install_fake_open_pr_gh_response(
-		&temp_dir, &worktree, pr_url, &head_oid,
-	);
-
-	tests::seed_review_handoff_marker_value(
-		&state_store,
-		config.service_id(),
-		&issue.id,
-		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
-	);
-
-	state_store
-		.record_run_attempt(run_id, &issue.id, 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease(config.service_id(), &issue.id, run_id, &issue.state.name)
-		.expect("stale lease should record");
-	state_store
-		.upsert_worktree(
-			config.service_id(),
-			&issue.id,
-			&worktree.branch_name,
-			&worktree.path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, false)
-		.expect("startup reconciliation should succeed");
-
-	assert!(
-		summary.is_none(),
-		"completed retained closeout with an open PR should stay blocked during startup recovery"
-	);
-	assert!(
-		state_store.lease_for_issue(&issue.id).expect("lease lookup should succeed").is_none(),
-		"startup reconciliation should clear stale completed closeout leases when the PR is still open"
-	);
-	assert_eq!(
-		state_store
-			.run_attempt(run_id)
-			.expect("run attempt lookup should succeed")
-			.expect("run attempt should still exist")
-			.status(),
-		"interrupted"
-	);
-	assert!(
-		state_store
-			.worktree_for_issue(&issue.id)
-			.expect("worktree lookup should succeed")
-			.is_some(),
-		"startup reconciliation should preserve the retained closeout worktree mapping until the PR merges"
-	);
-	assert!(
-		worktree.path.exists(),
-		"startup reconciliation should leave the retained closeout worktree on disk while waiting for merge"
-	);
-}
+mod closeout_leases;
+mod guarded_worktrees;
+mod terminal_labels;
